final updates, make image upload work
