+++ src/App.tsx
@@ -1,26 +1,8 @@
-import { useYDoc } from "zustand-yjs";
-import { DraggableLetter } from "./components/Letter";
 import people from "./data/people.json";
-import { IndexeddbPersistence } from "y-indexeddb";
 import * as Dialog from "@radix-ui/react-dialog";
-import { YJS_ROOM } from "./constants";
-import { LetterType, LetterSharedData, LetterInterface } from "./types";
-import { Letters } from "./data/letters";
-import { useEffect, useState } from "react";
-import {
-  LetterFormButton,
-  LetterFormDialogContent,
-} from "./components/LetterForm";
-import { ShuffleIcon, ResetIcon, ViewGridIcon } from "@radix-ui/react-icons";
-
-const connectDoc = (doc: Y.Doc) => {
-  console.log(`connected to ${doc.guid}`);
-  const index = new IndexeddbPersistence("(we)bsite-persistence", doc);
-  return () => {
-    index.destroy();
-    console.log("disconnected");
-  };
-};
+import { LetterType, LetterInterface } from "./types";
+import { LetterFormDialogContent } from "./components/LetterForm";
+import { Desk } from "./components/Desk";
 
 export const SubmitLetterMetadata: LetterInterface = {
   id: "submit-0",
@@ -37,20 +19,6 @@ export const SubmitLetterMetadata: LetterInterface = {
 };
 
 function App() {
-  const yDoc = useYDoc(YJS_ROOM, connectDoc);
-  const sharedMap = yDoc.getMap<LetterSharedData>("shared");
-  const [letters, setLetters] = useState<LetterInterface[] | undefined>();
-
-  useEffect(() => {
-    setLetters([
-      ...Letters,
-      {
-        ...SubmitLetterMetadata,
-        ctaContent: <LetterFormButton letter={Letters[0]}></LetterFormButton>,
-      },
-    ]);
-  }, []);
-
   return (
     <>
       <Dialog.Root>
@@ -82,29 +50,7 @@ function App() {
             <em>What do you want from the internet?</em>
           </p>
         </div>
-        <div className="toolbar">
-          {/* reset letters */}
-          <button>
-            <ResetIcon />
-          </button>
-          {/* shuffle letters */}
-          <button>
-            <ShuffleIcon />
-          </button>
-          {/* toggle between StackIcon */}
-          <button>
-            <ViewGridIcon />
-          </button>
-        </div>
-        <div id="desk">
-          {letters?.map((letter) => (
-            <DraggableLetter
-              letter={letter}
-              key={letter.id}
-              shared={sharedMap}
-            />
-          ))}
-        </div>
+        <Desk />
         <footer>
           (we)bsite is a project by{" "}
           <a href={people.spencer.url}>{people.spencer.fullName}</a> and{" "}
