+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const YJS_ROOM = "(we)bsite-persistence";
+export const FORM_STORAGE_KEY = "letter-form";
